# kin-kernel-20slot/service/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Kin Gateway v2 API
  version: 0.1.0
  description: Reference data-plane and control-plane contract.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/kernels:
    servers:
      - url: http://localhost:9090
    get:
      operationId: listKernels
      security: [{InternalBearer: []}]
      responses:
        "200": {description: Registered kernels.}
    post:
      operationId: registerKernel
      security: [{InternalBearer: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KernelRegistration"
      responses:
        "200": {description: Kernel registered.}
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/kernels/{id}/heartbeat:
    servers:
      - url: http://localhost:9090
    post:
      operationId: heartbeatKernel
      security: [{InternalBearer: []}]
      parameters:
        - $ref: "#/components/parameters/KernelID"
      responses:
        "204": {description: Heartbeat accepted.}
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/kernels/{id}/drain:
    servers:
      - url: http://localhost:9090
    post:
      operationId: drainKernel
      security: [{InternalBearer: []}]
      parameters:
        - $ref: "#/components/parameters/KernelID"
      responses:
        "202": {description: Drain requested.}
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/route-policies/{name}:
    servers:
      - url: http://localhost:9090
    get:
      operationId: getRoutePolicy
      security: [{InternalBearer: []}]
      parameters:
        - $ref: "#/components/parameters/PolicyName"
      responses:
        "200":
          description: Route policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutePolicy"
        "404":
          $ref: "#/components/responses/Error"
    put:
      operationId: putRoutePolicy
      security: [{InternalBearer: []}]
      parameters:
        - $ref: "#/components/parameters/PolicyName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutePolicy"
      responses:
        "200":
          description: Route policy stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutePolicy"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/snapshots/current:
    servers:
      - url: http://localhost:9090
    get:
      operationId: getCurrentSnapshot
      security: [{InternalBearer: []}]
      responses:
        "200": {description: Current immutable configuration snapshot.}
  /api/v1/reconcile:
    servers:
      - url: http://localhost:9090
    post:
      operationId: reconcileNow
      security: [{InternalBearer: []}]
      responses:
        "200": {description: Reconcile result.}
  /api/v1/config/routing:
    servers: [{url: http://localhost:9090}]
    get:
      operationId: getRoutingConfig
      security: [{InternalBearer: []}]
      responses:
        "200":
          description: Typed routing configuration and apply state.
          content: {application/json: {schema: {$ref: "#/components/schemas/RoutingEnvelope"}}}
    put:
      operationId: putRoutingConfig
      security: [{InternalBearer: []}]
      parameters: [{$ref: "#/components/parameters/OperatorHeader"}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/RoutingPut"}}}
      responses:
        "200":
          description: Routing configuration committed.
          content: {application/json: {schema: {$ref: "#/components/schemas/RoutingEnvelope"}}}
        "400": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /api/v1/config/model-policy:
    servers: [{url: http://localhost:9090}]
    get:
      operationId: getModelPolicyConfig
      security: [{InternalBearer: []}]
      responses:
        "200":
          description: Typed model policy and apply state.
          content: {application/json: {schema: {$ref: "#/components/schemas/ModelPolicyEnvelope"}}}
        "404": {$ref: "#/components/responses/Error"}
    put:
      operationId: putModelPolicyConfig
      security: [{InternalBearer: []}]
      parameters: [{$ref: "#/components/parameters/OperatorHeader"}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/ModelPolicyPut"}}}
      responses:
        "200":
          description: Model policy committed and effective.
          content: {application/json: {schema: {$ref: "#/components/schemas/ModelPolicyEnvelope"}}}
        "400": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /api/v1/slots/{id}:
    servers: [{url: http://localhost:9090}]
    parameters:
      - {name: id, in: path, required: true, schema: {type: string, maxLength: 128}}
    get:
      operationId: getSlotConfig
      security: [{InternalBearer: []}]
      responses:
        "200":
          description: Slot overrides, effective policy, and apply state.
          content: {application/json: {schema: {$ref: "#/components/schemas/SlotEnvelope"}}}
    patch:
      operationId: patchSlotConfig
      security: [{InternalBearer: []}]
      parameters: [{$ref: "#/components/parameters/OperatorHeader"}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/SlotPatch"}}}
      responses:
        "200":
          description: Slot desired policy committed.
          content: {application/json: {schema: {$ref: "#/components/schemas/SlotEnvelope"}}}
        "400": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /api/v1/slots/policy:
    servers: [{url: http://localhost:9090}]
    post:
      operationId: patchSlotPolicyBatch
      security: [{InternalBearer: []}]
      parameters: [{$ref: "#/components/parameters/OperatorHeader"}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/SlotBatchPatch"}}}
      responses:
        "200":
          description: Per-slot updates; commits are ordered and may be partial on a later failure.
          content: {application/json: {schema: {$ref: "#/components/schemas/SlotBatchResult"}}}
        "400": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /api/v1/config/proxy-pool:
    servers: [{url: http://localhost:9090}]
    get:
      operationId: getProxyPoolConfig
      security: [{InternalBearer: []}]
      responses:
        "200":
          description: Typed proxy pool config. GET never includes credentials.
          content: {application/json: {schema: {$ref: "#/components/schemas/ProxyPoolEnvelope"}}}
    put:
      operationId: putProxyPoolConfig
      security: [{InternalBearer: []}]
      parameters: [{$ref: "#/components/parameters/OperatorHeader"}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/ProxyPoolPut"}}}
      responses:
        "200": {description: Proxy pool config committed.}
        "400": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
  /api/v1/proxies:
    servers: [{url: http://localhost:9090}]
    get:
      operationId: listProxies
      security: [{InternalBearer: []}]
      responses:
        "200":
          description: Desensitized proxy list plus pool config and totals.
          content: {application/json: {schema: {$ref: "#/components/schemas/ProxyPoolSnapshot"}}}
  /api/v1/proxies/{id}/reveal:
    servers: [{url: http://localhost:9090}]
    parameters:
      - {name: id, in: path, required: true, schema: {type: string}}
    post:
      operationId: revealProxy
      security: [{InternalBearer: []}]
      parameters: [{$ref: "#/components/parameters/OperatorHeader"}]
      responses:
        "200":
          description: The only plaintext credential response. URI only. Do not cache.
          headers:
            Cache-Control: {schema: {type: string, enum: [no-store]}}
          content:
            application/json:
              schema:
                type: object
                required: [id, uri]
                additionalProperties: false
                properties:
                  id: {type: string}
                  uri: {type: string}
        "503": {$ref: "#/components/responses/Error"}
  /api/v1/config/documents/{domain}:
    servers:
      - url: http://localhost:9090
    parameters:
      - name: domain
        in: path
        required: true
        description: Configuration domain; may include slash-separated segments.
        schema: {type: string, maxLength: 128}
    get:
      operationId: getControlDocument
      security: [{InternalBearer: []}]
      responses:
        "200":
          description: Current valid document, or the last-known-good version with degraded=true.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ControlDocument"}
        "404": {$ref: "#/components/responses/Error"}
    put:
      operationId: putControlDocument
      security: [{InternalBearer: []}]
      parameters:
        - $ref: "#/components/parameters/OperatorHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ControlDocumentPut"}
      responses:
        "200":
          description: Versioned document committed.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ControlDocument"}
        "400": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
...
  /api/v1/operations:
    servers:
      - url: http://localhost:9090
    post:
      operationId: enqueueControlOperation
      security: [{InternalBearer: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/OperationEnqueue"}
      responses:
        "202":
          description: Operation enqueued or existing idempotent operation returned.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ControlOperation"}
        "400": {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    InternalBearer:
      type: http
      scheme: bearer
      description: KIN_CONTROL_INTERNAL_TOKEN shared only with trusted internal callers.
  parameters:
...
    KernelID:
      name: id
      in: path
      required: true
      schema: {type: string, maxLength: 128}
    PolicyName:
      name: name
      in: path
      required: true
      schema: {type: string, maxLength: 128}
    OperatorHeader:
      name: x-kin-operator
      in: header
      required: false
      schema: {type: string, maxLength: 128}
      description: Trusted operator identity supplied by the Node BFF; defaults to control-api.
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    KernelRegistration:
      type: object
      required: [id, address, capacity, provider]
      properties:
        id: {type: string}
        address: {type: string}
        capacity: {type: integer, minimum: 1}
        provider: {type: string}
        revision: {type: integer, minimum: 0}
    RoutePolicy:
      type: object
      required: [name, tenant, model_pattern, provider, slot_group, max_inflight, max_waiting_tool]
      properties:
        name: {type: string}
        tenant: {type: string}
        model_pattern: {type: string}
        provider: {type: string}
        slot_group: {type: string}
        max_inflight: {type: integer, minimum: 1}
        max_waiting_tool: {type: integer, minimum: 0}
...